fixing onboarding TL
