+++ corpan/corpan-app/package.json
@@ -1,7 +1,7 @@
 {
   "name": "corpan",
   "private": true,
-  "version": "0.18.0",
+  "version": "0.18.1",
   "type": "module",
   "scripts": {
     "dev": "vite",
