+++ corpan/corpan-app/src-tauri/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "corpan"
-version = "0.18.0"
+version = "0.18.1"
 description = "Language learning app"
 authors = ["you"]
 edition = "2021"
