+++ corpan/corpan-app/src-tauri/tauri.conf.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://schema.tauri.app/config/2",
   "productName": "corpan",
-  "version": "0.18.0",
+  "version": "0.18.1",
   "identifier": "com.corpora.corpan",
   "build": {
     "beforeDevCommand": "npm run dev",
