+++ corpan/corpan-app/src/store/translations.ts
@@ -34,7 +34,10 @@ export const WELCOME_BY_LANG = {
     th: "ยินดีต้อนรับ",
     vi: "Chào mừng",
     id: "Selamat datang",
+    jv: "Sugeng rawuh",
+    su: "Wilujeng sumping",
     ms: "Selamat datang",
+    tl: "Maligayang pagdating",
     sw: "Karibu",
     "zh-Hans": "欢迎",
     "zh-Hant": "歡迎",
@@ -96,7 +99,10 @@ export const AUTONYM_BY_LANG = {
     th: "ไทย",
     vi: "Tiếng Việt",
     id: "Bahasa Indonesia",
+    jv: "Basa Jawa",
+    su: "Basa Sunda",
     ms: "Bahasa Melayu",
+    tl: "Tagalog",
     sw: "Kiswahili",
     "zh-Hans": "中文（简体）",
     "zh-Hant": "中文（繁體）",
@@ -173,7 +179,10 @@ export const MAKE_PRIMARY_BY_LANG = {
     th: "ตั้งค่าไทยเป็นภาษาหลักของฉัน",
     vi: "Đặt tiếng Việt làm ngôn ngữ chính của tôi",
     id: "Jadikan Bahasa Indonesia bahasa utama saya",
+    jv: "Dadekake basa Jawa minangka basa utamaku",
+    su: "Jadikeun basa Sunda jadi basa utama abdi",
     ms: "Jadikan Bahasa Melayu sebagai bahasa utama saya",
+    tl: "Gawing pangunahing wika ko ang Tagalog",
     sw: "Fanya Kiswahili kuwa lugha yangu kuu",
     "zh-Hans": "将简体中文设为我的主要语言",
     "zh-Hant": "將繁體中文設為我的主要語言",
